move functions into AttachmentElementState
